feat(ansible): change community.general.proxmox_* to community.proxmox
Proxmox modules are being moved out of `community.general` and into
their own namespace.  There's a bit of a complication in that until now
we've used the `proxmoxer` module from the Debian repositories but this
only provides version 1.2.0 and the latest versions require at least
2.0.0.  So, we will create a Python virtual environment on the Proxmox
host for used by Ansible and install `proxmoxer` 2.2.0 (which is
currently the latest version).  Update `requirements.yml` to include
`community.proxmox==1.5.0`.